pyproject.toml
[project]
name = "gong-to-notion"
version = "0.1.0"
description = "Deterministic Gong call → Notion Customer Interactions importer."
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "python-dotenv>=1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/gong_to_notion"]
